doctor-appointment-api/requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0
aiomysql
python-jose
passlib[bcrypt]
pydantic
pytest
httpx
